Extract search component on web
