+++ web/app/components/OhmSearch.vue
@@ -1,0 +1,45 @@
+<template>
+    <div>
+      <form @submit.prevent="query">
+        <input id="tracking" type="text" v-model="trackingId" placeholder="Enter your tracking id"/>
+        <button type="submit">Submit</button>
+      </form>
+  </div>
+</template>
+
+<script>
+import axios from 'axios';
+
+export default {
+  data() {
+    return {
+      trackingId: null,
+    };
+  },
+  methods: {
+      async query() {
+        try {
+          const { data } = await axios.get(`/api/ohms/${this.trackingId}`);
+          this.$emit('found', data);
+        } catch (error) {
+          this.$emit('error', error.response && error.response.data);
+        }
+      },
+  }
+};
+</script>
+
+<style scoped>
+form {
+  text-align: center;
+
+}
+button, input {
+  display: block;
+  margin: 0 auto;
+}
+input {
+    margin: 2em auto 1em;
+    width: 50%;
+}
+</style>
